break(router): move `Bindings` to `worktop/cfw` module
