+++ packages/worktop/src/durable.ts
@@ -1,7 +1,7 @@
 import { connect } from 'worktop/ws';
-import type { Bindings } from 'worktop';
-import type { WebSocket } from 'worktop/ws';
+import type { Bindings } from 'worktop/cfw';
 import type { Durable } from 'worktop/durable';
+import type { WebSocket } from 'worktop/ws';
 
 export abstract class Actor {
 	DEBUG: boolean;
