+++ packages/worktop/types/router.strict.ts
@@ -2,7 +2,8 @@ import { Router } from 'worktop';
 import * as modules from 'worktop/cfw';
 import * as sw from 'worktop/sw';
 
-import type { Bindings, Context } from 'worktop';
+import type { Context } from 'worktop';
+import type { Bindings } from 'worktop/cfw';
 import type { Durable } from 'worktop/durable';
 import type { KV } from 'worktop/kv';
